Move AnalyticsProvider
